feat: report form

+++ src/main/java/com/timekiller/zzatool/test/service/AdminService.java
@@ -6,14 +6,14 @@ import com.timekiller.zzatool.test.dto.TestDTO;
 public interface AdminService {
 
     /* 관리자의 테스트 수정 */
-    void modifyTest(Integer testId, TestDTO testDTO);
+    void modifyTest(Long testId, TestDTO testDTO);
 
     /* 관리자의 테스트 삭제 */
-    void removeTest(Integer testId);
+    void removeTest(Long testId);
 
     /* 관리자의 퀴즈 수정 */
-    void modifyQuiz(Integer quizId, QuizDTO quizDTO);
+    void modifyQuiz(Long quizId, QuizDTO quizDTO);
 
     /* 관리자의 퀴즈 삭제 */
-    void removeQuiz(Integer quizId);
+    void removeQuiz(Long quizId);
 }
